+++ src/main/java/com/msg/after_school/domain/after_school/data/response/AfterSchoolResponse.java
@@ -8,9 +8,9 @@ import java.util.List;
 @Getter
 @Builder
 public class AfterSchoolResponse {
-    private Long id;
+    private Integer id;
     private String title;
     private List<String> week;
-    private List<Long> grade;
+    private List<Integer> grade;
     private Boolean isOpened;
 }
